refactor: add class User and moved interfaces
