+++ semana20/labook/.lintstagedrc.json
@@ -1,5 +1,5 @@
 {
   "*.ts": [
-    "eslint . --fix"
+    "tsc && eslint \"src/**/*.{ts}\" --quiet --fix"
   ]
 }
